flight_fighter_api_settings/requirements.txt
asgiref==3.2.10
asn1crypto==0.24.0
attrs==20.2.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==20.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.1.1
djangorestframework==3.11.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==2.0.0
incremental==16.10.1
iniconfig==1.0.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
packaging==20.4
PAM==0.4.2
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
pytest==6.1.0
python-apt==1.6.5+ubuntu0.3
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
sqlparse==0.3.1
ssh-import-id==5.7
systemd-python==234
toml==0.10.1
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.1.0
zipp==3.2.0
zope.interface==4.3.2
